+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 * Honor `Retry-After` header in `Retry` middleware.  The response will
   not be retried until the maximum of the backoff strategy and any
   time specified by the `Retry-After` header of the response.
+* The `RetryPolicy.defaultRetriable` only works for methods guaranteed
+  to not have a body.  In fs2, we can't introspect the stream to
+  guarantee that it can be rerun.  To retry requests for idempotent
+  request methods, use `RetryPolicy.unsafeRetriable`.  To retry
+  requests regardless of method, use
+  `RetryPolicy.recklesslyRetriable`.
+
+# v0.16.0-SNAPSHOT
+* `Retry` middleware takes a `RetryPolicy` instead of a backoff
+  strategy.  A `RetryPolicy` is a function of the request, the
+  response, and the number of attempts.  Wrap the previous `backoff`
+  in `RetryPolicy {}` for compatible behavior.
 
 # v0.17.0-RC3 (2017-08-29)
 * In blaze-server, when doing chunked transfer encoding, flush the